# Conflicts:
#	Sources/CMake/Modules/FindQT5.cmake
#	Sources/CMake/Modules/PlatformSettings.cmake
#	Sources/Internal/CMakeLists.txt
#	Sources/Internal/Render/RHI/GLES2/rhi_TextureGLES2.cpp
#	Tools/ResourceEditor/CMakeLists.txt


Former-commit-id: 12abbac915bc1e91344d1c2d4f76b37bb46ed2f7

+++ Sources/Internal/Render/RHI/DX9/rhi_QueryBufferDX9.cpp
@@ -96,7 +96,7 @@ dx9_QueryBuffer_Delete(Handle handle)
             cmd.push_back(c);
         }
 
-        ExecDX9(&cmd[0], cmd.size());
+        ExecDX9(&cmd[0], static_cast<uint32>(cmd.size()));
         buf->query.clear();
     }
